requirements.txt
google-genai>=1.0.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
ffmpeg-python>=0.2.0
numpy>=1.21.0